move search schemas into /api/schema
