+++ app/api/schema/search/facets/route.tsx
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import searchFacetsSchema from "../../../src/data/schemas/searchFacetsSchema";
+import searchFacetsSchema from "../../../../src/data/schemas/searchFacetsSchema";
 
 export const GET = async () => {
   return NextResponse.json(searchFacetsSchema, { status: 200 });
